Fix GELFRabbitHandler example
The `%2F` in the string must be a legacy approach.  If the string includes `%2F`, the vhost always ends up as `/`, ignoring the `virtual_host` kwarg.  I also made the `exchange` into an explicit kwarg.  It duplicates the default so it could actually be excluded, but making it explicit will help the user realize where the messages are found.